trying to fix differences to Frederik's work

+++ src/convert_era_5_to_seperate_basins.py
@@ -125,7 +125,7 @@ def parse_single_basin_precipitation(station_id, basin_data, discharge_file_name
             ind_lat_min = np.squeeze(np.argwhere(lat == max(min_lat, min_lat_array)))
             ind_lat_max = np.squeeze(np.argwhere(lat == min(max_lat, max_lat_array)))
             started_reading_data = True
-        tp = np.asarray(dataset['tp'][:, ind_lat_max:ind_lat_min + 1, ind_lon_min:ind_lon_max + 1])
+        tp = np.asarray(dataset['tp'][:, ind_lat_max:ind_lat_min + 1, ind_lon_min:ind_lon_max + 1], dtype=np.float32)
         tp[tp < 0] = 0
         # convert the time to datetime format and append it to the times array
         times = [datetime.datetime.strptime("1900-01-01 00:00", "%Y-%m-%d %H:%M") + datetime.timedelta(hours=int(ti[i]))
